CORE-4996 move mocks/DigestServiceMock into test/
