+++ ledger-obsolete/src/test/java/net/corda/v5/ledger/obsolete/merkle/MerkleTreeJavaApiTest.java
@@ -4,7 +4,7 @@ import net.corda.v5.crypto.DigestAlgorithmName;
 import net.corda.v5.crypto.DigestService;
 import net.corda.v5.crypto.DigestServiceUtils;
 import net.corda.v5.crypto.SecureHash;
-import net.corda.v5.ledger.obsolete.crypto.mocks.DigestServiceMock;
+import net.corda.v5.ledger.obsolete.mocks.DigestServiceMock;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
