+++ ledger-obsolete/src/test/kotlin/net/corda/v5/ledger/obsolete/merkle/MerkleTreeTests.kt
@@ -5,7 +5,7 @@ import net.corda.v5.crypto.DigestService
 import net.corda.v5.crypto.SecureHash
 import net.corda.v5.crypto.getZeroHash
 import net.corda.v5.crypto.create
-import net.corda.v5.ledger.obsolete.crypto.mocks.DigestServiceMock
+import net.corda.v5.ledger.obsolete.mocks.DigestServiceMock
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertNotEquals
 import org.junit.jupiter.api.Assertions.assertTrue
