+++ ledger-obsolete/src/test/kotlin/net/corda/v5/ledger/obsolete/merkle/PartialMerkleTreeTests.kt
@@ -3,7 +3,7 @@ package net.corda.v5.ledger.obsolete.merkle
 import net.corda.v5.crypto.DigestAlgorithmName
 import net.corda.v5.crypto.DigestService
 import net.corda.v5.crypto.SecureHash
-import net.corda.v5.ledger.obsolete.crypto.mocks.DigestServiceMock
+import net.corda.v5.ledger.obsolete.mocks.DigestServiceMock
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertFalse
 import org.junit.jupiter.api.Assertions.assertTrue
